+++ tests/test_contextmanager_py36.py
@@ -2,25 +2,25 @@ import asyncio
 
 import pytest
 
-from asyncio_extras import yield_async, async_contextmanager
+from asyncio_extras import async_contextmanager
 
 
 @pytest.mark.asyncio
 async def test_async_contextmanager():
     @async_contextmanager
     async def dummycontext(value):
-        await yield_async(value)
+        yield value
 
     async with dummycontext(2) as value:
         assert value == 2
 
 
 @pytest.mark.asyncio
-async def test_async_contextmanager_three_awaits():
+async def test_async_contextmanager_two_awaits():
     @async_contextmanager
     async def dummycontext(value):
         await asyncio.sleep(0.1)
-        await yield_async(value)
+        yield value
         await asyncio.sleep(0.1)
 
     async with dummycontext(2) as value:
@@ -28,24 +28,29 @@ async def test_async_contextmanager_three_awaits():
 
 
 @pytest.mark.asyncio
-async def test_async_contextmanager_no_yield():
+async def test_async_contextmanager_exception():
     @async_contextmanager
-    async def dummycontext():
-        pass
+    async def dummycontext(value):
+        nonlocal exception
+        try:
+            yield value
+        except Exception as e:
+            exception = e
 
-    with pytest.raises(RuntimeError) as exc:
-        async with dummycontext():
-            pass
+    exception = None
+    with pytest.raises(Exception):
+        async with dummycontext(2):
+            raise Exception('foo')
 
-    assert str(exc.value) == 'coroutine finished without yielding a value'
+    assert str(exception) == 'foo'
 
 
 @pytest.mark.asyncio
 async def test_async_contextmanager_extra_yield():
     @async_contextmanager
     async def dummycontext(value):
-        await yield_async(value)
-        await yield_async(3)
+        yield value
+        yield 3
 
     with pytest.raises(RuntimeError) as exc:
         async with dummycontext(2) as value: