+++ tests/test_generator_py36.py
@@ -1,0 +1,8 @@
+from asyncio_extras import isasyncgenfunction
+
+
+def test_isasyncgenfunction():
+    async def asyncgenfunc():
+        yield 1
+
+    assert isasyncgenfunction(asyncgenfunc)